Fixed process_board and bom_automator:
In bom_automator, I added a check for whether a cart already exists for a
distributor in process_board.

In distributor_cart, when we generate a part number CSV file or a URL CSV
file, the first row now labels the columns. Also, condensed the code for
truncating 0s in the URL in generate_url.